R package to support PEcAn, the Predictive Ecosystem Analyzer

Instructions for the use of this package are provided in the project documentation https://pecanproject.github.io/documentation.html.

Details

Project homepage: pecanproject.org

Description of PEcAn

The Predictive Ecosystem Analyzer (PEcAn) is a scientific workflow management tool that is designed to simplify the management of model parameterization, execution, and analysis. The goal of PEcAn is to streamline the interaction between data and models, and to improve the efficacy of scientific investigation. PEcAn is an open source utility that encapsulates:

acquisition of meteorological inputs

synthesis of physiological trait data as the posterior distribution of a Bayesian meta-analysis

sampling trait meta-analysis posterior distributions to parameterize ensembles of ED2 and other ecophysiological models

probabilistic forecasts

postprocessing to constrain forecasts and model parameters with field, meterological, eddy flux, and spectral data, and

provenance tracking

PECAn integrates available data into ecological forecasts by running ensembles of a terrestrial ecosystem model that is parameterized by the posterior distribution from a meta-analysis of available plant trait data. These trait data are assembled from field research and primary literature, and are stored in a PostgreSQL database. Current development focused on biofuel crops uses BETYdb. In addition to generating forecasts that reflect available data, PEcAn quantifies the contribution of each parameter to model uncertainty. This information informs targeted data collection and synthesis efforts that most efficiently reduce forecast uncertainty.

Current development is focused on developing PEcAn into a real-time data assimilation and forecasting system. This system will provide a detailed analysis of the past and present ecosystem functioning that seamlessly transitions into forecasts.

Author

Maintainer: Rob Kooper kooper@illinois.edu

Authors:

Mike Dietze dietze@bu.edu

David LeBauer dlebauer@email.arizona.edu

Xiaohui Feng feng22@illinois.edu

Dan Wang

Carl Davidson davids14@illinois.edu

Shawn Serbin sserbin@bnl.gov

Shashank Singh shashanksingh819@gmail.com

Chris Black chris@ckblack.org

Tanishq Jain tanishqjain010@gmail.com

Other contributors:

University of Illinois, NCSA [copyright holder]