Arrhenius scaling

Scale temperature dependent trait from measurement temperature to reference temperature

Usage

arrhenius.scaling(observed.value, old.temp, new.temp = 25)

Arguments

observed.value

observed value of temperature dependent trait, e.g. Vcmax, root respiration rate

old.temp

temperature at which measurement was taken or previously scaled to

new.temp

temperature to be scaled to, default = 25 C

Value

numeric value at reference temperature

Author

unknown