Get sensitivity analysis samples

Samples parameters for a model run at specified quantiles.

get.sa.samples(samples, quantiles)

Arguments

samples

random samples from trait distribution

quantiles

list of quantiles to at which to sample, set in settings file

Value

a list of lists representing quantile values of trait distributions

Details

Samples from long (>2000) vectors that represent random samples from a trait distribution. Samples are either the MCMC chains output from the Bayesian meta-analysis or are randomly sampled from the closed-form distribution of the parameter probability distribution function. The list is indexed first by trait, then by quantile.

Author

David LeBauer