Load an RData file into a list

Instead of polluting the current environment, this allows you to read an RData file into a list object of whatever name you choose.

Usage

load_local(file)

Arguments

file

a (readable binary-mode) connection or a character string giving the name of the file to load (when tilde expansion is done).

Value

List, with names corresponding to object names in file

Author

Alexey Shiklomanov

Examples

x <- 1:10
y <- 11:15
tmp <- tempfile()
save(x, y, file = tmp)
my_list <- load_local(tmp)
rm(tmp)