Run model specific write configuration functions

Main driver function to call the ecosystem model specific (e.g. ED, SiPNET) run and configuration file scripts

run.write.configs(settings, write = TRUE,
  ens.sample.method = "uniform", posterior.files = rep(NA,
  length(settings$pfts)), overwrite = TRUE)

Arguments

write
should the runs be written to the database

ens.sample.method
how to sample the ensemble members('halton' sequence or 'uniform' random)

posterior.files
Filenames for posteriors for drawing samples for ensemble and sensitivity analysis (e.g. post.distns.Rdata, or prior.distns.Rdata). Defaults to NA, in which case the most recent posterior or prior (in that order) for the workflow is used. Should be a vector, with one entry for each PFT. File name only; PFT outdirs will be appended (this forces use of only files within this workflow, to avoid confusion).

model
the ecosystem model to generate the configuration files for

Value

an updated settings list, which includes ensemble IDs for SA and ensemble analysis

Details

DEPRECATED: This function has been moved to the PEcAn.workflow package and will be removed from PEcAn.utils.