Standardized variable names and units for PEcAn

A lookup table giving standard names, units and descriptions for variables in PEcAn input/output files. Originally based on the MsTMIP standards, with additions to accomodate a wider range of model inputs and outputs. The standard_vars table replaces both mstmip_vars and mstmip_local, both of which are now deprecated.

standard_vars

Format

data frame, all columns character

Variable.Name

Short name suitable for programming with

standard_name

Name used in the NetCDF CF metadata conventions

Units

Standard units for this variable. Do not call variables by these names if they are in different units. See udunits for conversions to and from non-standard units

Long.Name

Human-readable variable name, suitable for e.g. axis labels

Category

What kind of variable is it? (Carbon pool, N flux, dimension, input driver, etc)

var_type

Storage type (character, integer, etc)

dim1,dim2,dim3,dim4

Dimensions across which is this variable allowed to vary. Dimension names are themselves standard vars and must be present in the table with category "Dimension"

Description

Further details. For composite measures, list the variables it is calculated from